requirements.txt
customtkinter==5.2.2
Pillow==10.1.0
numpy==1.24.3
